+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## [1.0.0]
+### Added
+- New startup activity which displays plugin info page with links and support section.
 ### Changed
 - Plugin Icon to a new, more suitable one.
 - Updated description to better explain a plugin